feat(refactor) - refactor the crypto package
